pyproject.toml
[project]
name = "infinite-chat"
version = "0.1.0"
description = "Minimal terminal chat client with infinite conversation history"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.119.0",
    "httpx>=0.28.1",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]

[dependency-groups]
dev = []
